Use iterate to go through all messages (now called payloads)
- Rename messages to payloads
  - message is a reserved light-service word
  - payload is also what we've been calling the whole string
  in other parts of the code
- Do not blow up if a channel is seen that is not being watched
