+++ lib/message_bus_client_worker.rb
@@ -6,6 +6,7 @@ require "securerandom"
 require "sidekiq"
 require "sidekiq-unique-jobs"
 require "active_support/core_ext/hash/indifferent_access"
+require "active_support/core_ext/string/inflections"
 require "message_bus_client_worker/version"
 require "message_bus_client_worker/workers/enqueuing_worker"
 require "message_bus_client_worker/workers/subscription_worker"
@@ -15,8 +16,8 @@ require "message_bus_client_worker/services/polling/set_last_id"
 require "message_bus_client_worker/services/polling/generate_client_id"
 require "message_bus_client_worker/services/polling/generate_uri"
 require "message_bus_client_worker/services/polling/generate_params"
-require "message_bus_client_worker/services/polling/get_messages"
-require "message_bus_client_worker/services/polling/process_messages"
+require "message_bus_client_worker/services/polling/get_payloads"
+require "message_bus_client_worker/services/polling/process_payload"
 require "message_bus_client_worker/services/poll"
 
 module MessageBusClientWorker
