+++ message_bus_client_worker.gemspec
@@ -36,11 +36,11 @@ Gem::Specification.new do |spec|
   spec.add_dependency "addressable"
   spec.add_dependency "light-service"
   spec.add_dependency "sidekiq-unique-jobs", ">= 6.0.0"
+  spec.add_dependency "activesupport"
 
   spec.add_development_dependency "bundler", "~> 1.16"
   spec.add_development_dependency "rake", "~> 10.0"
   spec.add_development_dependency "rspec", "~> 3.0"
   spec.add_development_dependency "rspec-sidekiq"
-  spec.add_development_dependency "activesupport"
   spec.add_development_dependency "wait"
 end
